95 work

+++ src/main/stream/Constaints.java
@@ -2,6 +2,6 @@ package main.stream;
 
 public class Constaints {
 
-	public static int INFINITY = (int)Double.POSITIVE_INFINITY;
+	public static double INFINITY = Double.POSITIVE_INFINITY;
 	
 }
